+++ doc/CHANGELOG.md
@@ -9,6 +9,10 @@
 
 - Fixed some unexpected file saving. (#353)
 
+### Changed
+
+- Open an empty untitled tab when the open file length limit is exceeded. (#353)
+
 ## 6.4.1 (Beta)
 
 ### Linting with Language Server